pyproject.toml
[project]
name = "voice-chat-app"
version = "1.0.0"
description = "Voice Chat App mit Claude AI"
authors = [
    {name = "Voice Chat App", email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "numpy==2.2.6",
    "openai==1.82.0",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic_core==2.33.2",
    "pyperclip==1.9.0",
    "PyQt5==5.15.11",
    "PyQt5-Qt5==5.15.2",
    "PyQt5_sip==12.17.0",
    "pywin32==310",
    "scipy==1.15.3",
    "sniffio==1.3.1",
    "sounddevice==0.5.2",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.13.2",
    "winshell==0.6",
    "requests==2.32.3",
    "pygame==2.6.1",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.uv.workspace]
members = [
    "voice-chat-project",
]

[project.scripts]
voice-chat = "voice_chat_app:main"
